Moved chat code to server/src/chat
