+++ server/src/chat/entity2termObj.ts
