+++ server/src/chat/resource.ts
