+++ server/src/chat/route.ts
@@ -1,21 +1,21 @@
 import type { ChatRequest, ChatResponse, LlmConfig, QueryClassification, RouteApi, RoutePayload } from '#types'
 import { mayLog } from '#src/helpers.ts'
 import { formatElapsedTime } from '#shared'
-import { readJSONFile, parse_geneset_db, getChatRelatedPlotTypes } from './chat/utils.ts'
-import { classifyQuery } from './chat/classify1.ts'
-import { classifyPlotType } from './chat/plot.ts'
-import { classifyNotPlot } from './chat/classify2.ts'
-import { inferScaffold } from './chat/scaffold.ts'
+import { readJSONFile, parse_geneset_db, getChatRelatedPlotTypes } from './utils.ts'
+import { classifyQuery } from './classify1.ts'
+import { classifyPlotType } from './plot.ts'
+import { classifyNotPlot } from './classify2.ts'
+import { inferScaffold } from './scaffold.ts'
 import serverconfig from '#src/serverconfig.js'
 import { getDsAllowedTermTypes } from '../../routes/termdb.config.ts'
-import { phrase2entity } from './chat/phrase2entity.ts'
-import { inferTermObjFromEntity } from './chat/entity2termObj.ts'
-import { resolveToTwTvs } from './chat/entity2twTvs.ts'
-import { answerDataQueries } from './chat/dataQueries.ts'
+import { phrase2entity } from './phrase2entity.ts'
+import { inferTermObjFromEntity } from './entity2termObj.ts'
+import { resolveToTwTvs } from './entity2twTvs.ts'
+import { answerDataQueries } from './dataQueries.ts'
+import type { Scaffold, Phrase2EntityResult, SummaryScaffold } from './scaffoldTypes.ts'
+import { resolveToPlotState } from './scaffold2state.ts'
 import path from 'path'
 import fs from 'fs'
-import type { Scaffold, Phrase2EntityResult, SummaryScaffold } from './chat/scaffoldTypes.ts'
-import { resolveToPlotState } from './chat/scaffold2state.ts'
 
 const payload: RoutePayload = {
 	init,
