+++ server/src/chat/scaffold.ts
