+++ server/src/chat/scaffold2state.ts
