+++ server/src/chat/scaffoldTypes.ts
