+++ server/src/chat/utils.ts
@@ -5,7 +5,7 @@ import { route_to_appropriate_llm_provider } from './routeAPIcall.ts'
 import type { MsgToUser, Entity, FilterTreeNode, FilterLeafNode, FilterTreeResult } from './scaffoldTypes.ts'
 import { filterTreeJsonSchema } from './scaffoldTypes.ts'
 import fs from 'fs'
-import { getDsAllowedTermTypes } from '../../../routes/termdb.config.ts'
+import { getDsAllowedTermTypes } from '../../routes/termdb.config.ts'
 import Database from 'better-sqlite3'
 import { GENE_FEATURE_KEYWORDS, determineAmbiguousGenePrompt } from './determineAmbiguousGene.ts'
 import { GENE_SET_KEYWORDS } from './genesetdatatype.ts'
