+++ server/src/test/routes/testchat.ts
@@ -3,8 +3,8 @@
 // Test URL: http://localhost:3000/testchat
 import serverconfig from '../../serverconfig.js'
 import { getDsAllowedTermTypes } from '../../../routes/termdb.config.ts'
-import { readJSONFile, getChatRelatedPlotTypes } from '../../routes/chat/utils.ts' // getGenesetNames
-import { run_chat_pipeline } from '#src/routes/termdb.chat.ts'
+import { readJSONFile, getChatRelatedPlotTypes } from '../../chat/utils.ts' // getGenesetNames
+import { run_chat_pipeline } from '#src/chat/route.ts'
 import assert from 'node:assert/strict'
 import path from 'path'
 import fs from 'fs'
